chore(ci): Add the dependency packages to the DEB control file
Also, it turns out that for some compilers, the static linking in
Release mode does not work across libraries (e.g. monomuxServer could
not link monomuxControl), so we had to refactor to have a single core
library (that can be unit-tested and linked against) and an executable
that contains the `main()` things.