fix(UB): Client callback handler crashes/segfaults
The `setupDispatch()` routine both in the `Client` and the `Server` set
up a `std::bind()` of their respective `this`, but the `Client`, when
handling a message, started reading bogus data...

The culprit was that during normal operation, the `Client` is moved-from
when switching from `src/main` to `src/Client/Main`. The server
contained the same bug, but because the `Server` is never moved, it did
not appear.