fix: Wake up for already buffered reads
The previous commit introduced more tough local buffering for data to
prevent lock-ups in case the kernel-level buffers are full. However,
locally buffered data is not part of the file descriptor anymore, so `epoll`
cannot signal the availability of data.

Use `eventfd` and implement a hand-made re-schedule call to alleviate
this.