feat: Kick clients that misbehave and consume data too slowly
This is another step towards trying to make the server more robust. If
the data connection with a client overflows it means they stopped
communicating. However, large outputs by sessions being sent to clients
would result in the server process continuing to chew up memory...

A defence mechanism is implemented which kicks the clients if they
overflow a specific buffer size.