chore(licence)!: Clarify the licensing situation
First, there are parts of MonoMux which are designed to be useful as a
reusable library. This is generally prevented by a blanket GPL clause,
and as such the post-refactoring version will offer these library
elements under the LGPL instead.
The user-facing application parts will remain licensed under the GPL.

Second, the new year has dawned and this would require updating the
copyright year timestamp every time, which would result in crappy
commits polluting the history of the entire project (just like how this
commit in particular does so!). It seems that the world has moved on
from having 17 lines of comments prefacing every file to
[machine-readable identifiers][1].
If several famous and large projects [2][3] can get away with simply
including the licence metadata without the boilerplate, surely we can,
to...

[1]: http://spdx.org/licenses/
[2]: http://github.com/llvm/llvm-project/blob/llvmorg-14.0.0/clang-tools-extra/clang-tidy/tool/ClangTidyMain.cpp#L5
[3]: http://www.kernel.org/doc/html/v6.0/process/license-rules.html