refactor(dto-compiler): Do not instantiate nested `namespace`s in the DTO AST
Nested namespaces existing as a single entity will likely cause
significant extra processing requirements for lookup. Instead, flatten
the hierarchy, and instantiate every instance of the "same" namespace
separately, so that the AST is a more faithful representation of the
original input DSL. This requires the ability to chain `decl_context`s,
which was implemented in this patch.